--- src/del/layout.tsx
@@ -1,37 +1,0 @@
-import type { Metadata } from "next";
-import Header from "@/components/commons/Header";
-// フォントの読み取り
-import { M_PLUS_1 } from "next/font/google";
-import "@assets/css/reset.css";
-import "@assets/css/style.css";
-
-const title = "Bakery NIKORI";
-
-const mPlus1 = M_PLUS_1({
-  weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"], // 使用したい太さを指定
-  subsets: ["latin"], // サブセット（必要なら "latin-ext" や "japanese" など）
-  display: "swap", // 推奨設定
-});
-
-// メタデータの設定
-export const metadata: Metadata = {
-  title: title,
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="ja" className={mPlus1.className}>
-      <body>
-        <Header title={title} />
-        <main>
-          {children}
-        </main>
-      </body>
-    </html>
-  );
-}
