+++ src/pages/index.tsx
@@ -4,6 +4,10 @@ import "@assets/css/style.css";
 
 import Keyvisual from "@/components/section/Keyvisual";
 import Concept from "@/components/section/Concept";
+import Point from "@/components/section/Point";
+import BreadSec from "@/components/section/TopBread";
+import TopCafe from "@/components/section/TopCafe";
+import News from "@/components/section/News";
 
 
 
@@ -12,6 +16,13 @@ export default function Home() {
     <>
         <Layout>
             <Keyvisual/>
+            <div className="container">
+              <Concept/>
+              <Point/>
+              <BreadSec/>
+              <TopCafe />
+              <News/>
+            </div>
         </Layout>
     </>
   );
